// src/app/shared/models/documents/documents-request.model.ts
export interface DocumentsRequest {
  documentId: number;
  documentTypeId: number;
  documentTitle: string;
  description?: string | null;
  moduleName: string;
  referenceId: number;
  originalFileName?: string | null;
  storedFileName?: string | null;
  filePath?: string | null;
  fileExtension?: string | null;
  fileSizeKB?: number | null;
  versionNo?: number | null;
  effectiveDate?: string | null;
  expiryDate?: string | null;
  isConfidential?: boolean | null;
  status?: string | null;
  uploadedBy?: number | null;
  file?: File | null;   
}